Refactor google credentials into a shared function
